Divide class FelineTest on with parametrized and without parametrized
